Reconstruction - part 1
First part of the changes that introduces reconstruction support.
Reconstruction is the process of recreating memory allocator and garbage
collector metadata of the persistent memory region that was in-use
before application crash.
This commit adds code that walks the span bitmap log, figures out the
properties of the span to be reconstructed and calls the core
reconstruction function.