Add reconstructed span to appropriate allocator/GC lists
This commit sets the remaining metadata of the span, adds it to the
appropriate memory allocator list depending on the number of pages in
the span, and also marks it for garbage collection.
Added a new EnableGC() API that users need to call after setting the
persistent memory root pointer to enable GC and to clean up unused
memory in reconstructed memory regions.