Use persistent memory mapped address to set/get root
As the address at which persistent memory file gets mapped can change
across application run, the offset to the root pointer is stored in the
header region rather than the absolute address. Earlier, the offset was
computed with respect to pmemInfo.startAddr. But if done so, the offset
value can be 0. On the next GetRoot() call, the runtime cannot determine
if this implies the root is unset or the root corresponds to
pmemInfo.startAddr. Hence this commit uses pmemInfo.mappedAddr for root
get and set APIs.