Rename persistent as memtype
The name persistent was used as function parameter to indicate whether
memory has to be allocated from persistent memory or volatile memory.
But this can be confusing as it might be read as the function handles
only memory management related to persistent memory. Hence, rename
persistent as memtype in function parameters, comments, etc.