Bug fix in treap search code
Earlier, the heap traversal code recursively searched the right child
only when the number of pages in the right child was equal or greater
than the required number of pages. But this is incorrect.
The 'free' treap is a randomized binary search tree. Hence the number of
pages in the right child will always be greater than the number of pages
in the current node.