Clear heap type bits for noscan objects
When the allocator reuses memory regions to satisfy an allocation
request, the logged heap type bits in persistent memory need to be
cleared. This is because, on reinitialization, the runtime need to be
sure what regions contain pointers and what regions are static data in
order to do swizzling.
This commit adds code to clear heap type bits when a persistent memory
span gets reused.
Also fixes the following issues
- missing logHeapBits() call when the allocated object is pointer-sized
- incorrect computation of number of heap bytes to log in heapBitsSetType()