Run treapsearch on systemstack
treapsearch uses a recursive strategy to find the required span in the
treap datastructure. This can cause the goroutine to run out of stack
space and result in the runtime trying to allocate a bigger stack for
it. treapsearch is called with _mheap locked. So allocating a new stack
will deadlock as it tries to lock _mheap as well.
This commit changes treapsearch to be run on the systemstack which is
much larger than the goroutine stack.