Add an offset field to file map API
This commit adds an offset field to the mapFile() file map API.
This will enable internal parts of a file to be mapped.
It also fixes an issue where the device size of devdax file was
being incorrectly obtained using fstat. fstat does not give the size of
a character device. Instead, the size is now read from the sysfs
filesystem using the /sys/dev/char/M:N/size file, where M = major num
and N = minor num of the device.