Heap type bit logging
This commit adds support for heap type bit logging/clearing for
persistent memory allocations. The heap type bits are logged in the
per-arena metadata section. The type bits are used by the GC to identify
which regions in memory have pointers in them. These type bits will be
restored in the next 'reconstruction' phase.
Also fixes an issue where an incorrect constant was used to compute page
size.