Span bitmap logging
Adding support for span bitmap logging. The span bitmap captures the
metadata of each in-use persistent memory span. The span metadata is
written into the span bitmap in the arena header. When the span gets
freed, the metadata is cleared. This metadata is used to recreate spans
during the reconstruction phase.