Reconstruction: arena mapping
Added code that maps all arenas found in the persistent memory file and
calls the reconstruct() function to recreate spans found in the span
table.
Also fixes an error in the treap search code. The mheap 'free' treap
uses the number of pages in a span for its primary sorting. Hence the
recursive search on the left subtree has to be done when the number of
pages of the span we are looking for is less than or equal to the number
of pages in the current span.