Swizzle arenas in a per-span basis
Changed swizzle algorithm to go through arenas in a per-span basis. This
is similar to how GC scans for objects in the heap (scanobject() in
mgcmark.go). If a span is not is-use, then the entire span is skipped
for swizzling rather than go through each 8 byte region. Similarly, if
there are no pointers left in an object, the remaining region in the
object is skipped for swizzling.