Optimize span creation
On profiling PmemInit, it was found that the initialization function
spent a significant portion of time in treapSearch() function. The
earlier implementation used an ineffective way to recreate spans -
trimming out ununsed regions of the recreated span into the mheap free
treap, and then later recursively searching the treap to find the spans
that were inserted into it earlier.
This commit changes the span recreation algorithm to recreate spans in a
more optimized manner. reconstruct() function iterates through the span
bitmap and each contiguous unused region of memory is inserted into the
free treap as a free-span.
In-use spans are inserted into the mheap busy list or the mcentral lists
depending on the span size.