Remove devdax support
The current heap recovery mechanism used does not work on devdax devices
because devdax devices cannot be mmap'd in a per-arena basis. Instead
the entire size of the device should be mapped at once.
This commit removes support for devdax devices.
Most functions in pmemUtil.go are called from mallocgc when persistent
memory is requested for the first time by an application.  Hence this
commit also modifies those functions to ensure that it does not cause
further memory allocations as a side-effect which would result in a
malloc deadlock.