Redirect memory stores within txn() block to Log() method
Injects SSA nodes to call Log() function of transaction interface.
This is done for all assignment statements within the transaction
that trigger a memory store op. Stores to registers/known stack
variables are not routed to Log(), so there will be less calls to
Log() function.

There are some problems when escape analysis determines a variable
is on stack, but it is passed to Log() function. If the stack is
now reallocated, it could lead to invalid memory operations.