Swizzling algorithm implementation - part 1
Added first part of the swizzleArenas() functions that swizzles pointers
in a per-arena basis. Swizzling is the process of re-writing pointers in
the persistent memory file when the memory-mapping address of arenas
change on a subsequent run of the application.
This commit adds code to set up the offsetTable and rangeTable, and call
swizzle() on each arena.