Add runtime.inpmem check before each store within txn() block
With this change, tx.Log() is called during program run only if a pointer
is in pmem. This also allows us to have volatile stores within txn() block.
So, the logic to force escape variables within txn() block introduced in
a previous commit has been removed with this change. This change also uses
AST node OTXBLOCK to track stores. So the logic to track stores during AST
optimizations has been removed with this change.