Change buildssa to use per-goroutine tx ptr inside txn()
With this change, each goroutine has a pointer to store handle of
the ongoing transaction. So, function calls made from within txn()
block (and on the same goroutine) can continue in the same transaction.
Any function that has to perform updates transactionally must have a
txn() block.
The per-goroutine transaction handle is set to nil, when the outermost
transaction is over. With this change, new tx, tx begin, end & tx release
are all inserted during buildssa. So, this change also removes this code
injection logic from parser.