Change where txn flag is added to go build
Previously, it was added to forcedGcFlags, which triggered the
re-compilation of a lot of Go compiler files with the -txn flag.
This made the compilation slower by ~8x for go-redis. With this
change, the compilation time is slowed down by ~1.5x. The baseline
is compilation of go-redis using transaction package and no txn-related
changes to go compiler.