Change ssa to check for bool returned from tx.End()
Before this commit, buildssa assumed that tx.End() would return
``error` type, and it used this value to figure out if a
transaction handle can be released. Creating a new Error type and
returning it is also a performance bottleneck as illustrated when
I changed test/bench/go1/binarytree_test.go and wrapped the code
with "txn() block". This caused ~2x slowdown even though there is
no activity in pmem. This test stresses memory allocator and GC.
End's errors.New() came up as a bottleneck in profiling.

After this commit, buildssa assumes tx.End() returns a bool. With
this, slowdown for binarytree_test.go is now ~1.35x