Change buildssa to allow return from within txn()
Before this commit, code with return inside within txn() would not
compile.
After this commit, Return within txn() block are treated as if the
transaction completed successfully. Also add code to disable
explicit nesting of txn() block in the same function. We fail during
compilation if user does this. This was done to keep the semantics
of returning simple.
Open question if we allow nesting of txn() within same function:
Do all the levels of transaction complete, or only inner one completes?