Handle function literals with txn() usage.
Before this there was a bug. noder.go was marking which function
has txn block. If a function literal didn't have txn block,
the entire function was marked as not having txn block.
Functions need to be marked as having txn block, because buildSSA
uses this to track new/pnew/make/pmake allocations of variables
within function. This tracking needs to start before OTXBLOCK
is encountered!