Add an allocation profiling thread
First commit on adding support for optimized persistent memory
allocations. For pmem allocation of an object containing a pointer, each
allocation results in heap type bits being logged for that object. For a
datatype that is frequently allocated, repeated heap type bit logging
can be avoided by specially caching spans for each such datatype in the
mcache. Then, heap type bits need to be set only for the first
allocation from such a span.
This commit adds two functions - typeIndex() that accounts allocation
information for each type and typeProfileThread() that promotes types to
be specially cached when certain heuristics are met.