Implementation of CollectPtrs
To support implementation of Log3() logging function in the transaction
package, runtime exports a function CollectPtrs() that finds pointers in
an object. Since Log3() uses a byte array to log objects, keeping a copy
of pointers found in the object in a separate array helps the GC keep
those objects alive.