Extend mcache, mcentral by one more dimension for type caching
To support caching spans according to the type of frequently allocated
datatypes, mcache and mcentral data structures have now been extended by
one more dimension. The third dimension can now hold up to 50 cached
types. If a type has not yet been specially cached, they are allocated
from index 0. Each type that is promoted is assigned an index, and spans
of that type are cached in the corressponding index in the mcache and
mcentral.