Change heap type bit logging function
This commit introduces changes in how heap type bits are logged. For
spans that are specially cached for the allocation of a particular
datatype, heap type bits are logged only for the first object
allocation. The logged data representation is also different. For a
specially cached span, logHeapBits() logs type index and the type
metadata which consits of kind, size, ptrdata and 1-bit heap type bits.