Heap type bit recovery with optimized logging
Changed restoreSpanHeapBits() function implementation for restoring heap
type bits for those spans that are specially cached for serving
allocations of a particular datatype. For such spans, the heap type bits
are set only for the first object. So heapBitsSetType() is called for
restoring the type bits for each object separately.