Inject code to start/close the transaction on seeing txn() block
With this change, buildssa() will use the per goroutine pointer
to the transaction handle introduced in a previous commit.
When a txn() block is first seen in a function, we check if there
is an ongoing transaction already. If yes, we reuse the transaction
handle. This way, we handle nested transactions or function calls
within transactions. Note that this function should be in the same
goroutine and if this function needs to make transactional updates,
it should have a txn() block of its own.

The per-goroutine transaction handle is set to nil, when the outermost
transaction is over. With this change, new tx, tx begin, end & tx release
are all inserted during buildssa in gc/ssa.go. Return within txn() block
are treated as if the transaction completed successfully.