Instrument stores in txn() block and redirect them to undo log
This is done in logStore ssa pass introduced before.
Stores within txn() block are intercepted, checked to be inPmem() and
logged by transaction package in this pass. This pass runs before
"writebarrier" pass. Writebarrier pass updates OpStore to OpWB.
So, if logStore pass runs after writebarrier, we'll need to track OpWB
inside txn() block too! This pass also runs before "lower" pass. Lower
pass converts genericOps like OpStore, OpMove, OpZero to platform-specific
opcodes. So if logStore runs after lower pass we'll need to track a lot
of platform specific Ops within txn() block. All the ssa optimizations
before this pass run as is. This ssa pass is not run for functions
without a txn() block and if the code is not built with a
"-txn" build flag.