Don't add runtime.InPmem check for stores known to be outside pmem at compile time
If a datastructure is allocated in a function using pnew/pmake, escape
analysis will determine if it will be in pmem or not. This is known at compile
time. For updates to such datastructures within txn() block, we won't add
runtime.InPmem() checks. This will help preserve the structure of CFG. The
implementation is best-effort currently. If we can't determine a variable's
location at compile time, we will add runtime.InPmem() check before the
variable's update. No significant performance improvements after this change.