Speed up build time with txn flag
This commit does 2 things to speed up the build time of files
with txn flag:
1. Change the place where txn flag is added to go build.
Previously, it was added to forcedGcFlags, which triggered the
re-compilation of a lot of Go compiler files with the -txn flag.
This made the compilation slower by ~8x for go-redis.
2. Enables parallel compilation of Go files even with txn flag.
Go disables parallel compilation of files for most flags unless they
are specifically whitelisted. With this change "txn" flag is
one of the flags for which parallel compilation is enabled.

With these 2 changes, there is no significant slowdown with txn
flag in build time of codebases like go-redis.

This commit also has a cosmetic change so build flags remain
alphabetically sorted in cmd/go/internal/cfg/cfg.go