Enable a debug option to force swizzle arenas
Usually runtime is able to map arenas at the exact address as it was
mapped previously. This makes testing pointer swizzling code. This
commit adds a forced arena swizzle option. Enabling this option maps
each arena at an offset which is a multiple of 1GB, thereby forcing
runtime to swizzle every arena.